Revert "Added Ban Button, Clear Queue, Video notes"
