+++ package.json
@@ -35,11 +35,9 @@
     "react-dom": "17.0.2",
     "react-icons": "^4.3.1",
     "react-player": "^2.9.0",
-    "react-rewards": "^2.0.3",
     "react-toastify": "^8.2.0",
     "redis-om": "^0.2.0",
-    "swr": "^1.2.2",
-    "zustand": "^4.0.0-rc.0"
+    "swr": "^1.2.2"
   },
   "devDependencies": {
     "@babel/core": "^7.17.7",
